+++ lib/vagrant-lxc/driver/builder.rb
@@ -5,9 +5,9 @@ module Vagrant
   module LXC
     class Driver
       class Builder
-        def self.build(id)
-          version = CLI.new.version.match(/^(\d+\.\d+)\./)[1].to_f
-          Driver.new(id).tap do |driver|
+        def self.build(id, shell)
+          version = CLI.new(shell).version.match(/^(\d+\.\d+)\./)[1].to_f
+          Driver.new(id, shell).tap do |driver|
             mod = version >= 0.8 ?
               Driver::FetchIpWithAttach :
               Driver::FetchIpFromDsnmasq