+++ lib/vagrant-lxc/driver/cli.rb
@@ -17,12 +17,10 @@ module Vagrant
           end
         end
 
-        # Include this so we can use `Subprocess` more easily.
-        include Vagrant::Util::Retryable
-
-        def initialize(name = nil)
-          @name   = name
-          @logger = Log4r::Logger.new("vagrant::provider::lxc::container::cli")
+        def initialize(sudo_wrapper, name = nil)
+          @sudo_wrapper = sudo_wrapper
+          @name         = name
+          @logger       = Log4r::Logger.new("vagrant::provider::lxc::container::cli")
         end
 
         def list
@@ -66,7 +64,7 @@ module Vagrant
         end
 
         def start(overrides = [], extra_opts = [])
-          options = overrides.map { |key, value| ["-s", "lxc.#{key}=#{value}"] }.flatten
+          options = overrides.map { |key, value| ["-s", "lxc.#{key}='#{value}'"] }.flatten
           options += extra_opts if extra_opts
           run :start, '-d', '--name', @name, *options
         end
@@ -110,56 +108,7 @@ module Vagrant
         private
 
         def run(command, *args)
-          execute('sudo', "lxc-#{command}", *args)
-        end
-
-        # TODO: Review code below this line, it was pretty much a copy and
-        #       paste from VirtualBox base driver and has no tests
-        def execute(*command, &block)
-          # Get the options hash if it exists
-          opts = {}
-          opts = command.pop if command.last.is_a?(Hash)
-
-          tries = 0
-          tries = 3 if opts[:retryable]
-
-          sleep = opts.fetch(:sleep, 1)
-
-          # Variable to store our execution result
-          r = nil
-
-          retryable(:on => LXC::Errors::ExecuteError, :tries => tries, :sleep => sleep) do
-            # Execute the command
-            r = raw(*command, &block)
-
-            # If the command was a failure, then raise an exception that is
-            # nicely handled by Vagrant.
-            if r.exit_code != 0
-              if @interrupted
-                @logger.info("Exit code != 0, but interrupted. Ignoring.")
-              else
-                raise LXC::Errors::ExecuteError, :command => command.inspect
-              end
-            end
-          end
-
-          # Return the output, making sure to replace any Windows-style
-          # newlines with Unix-style.
-          r.stdout.gsub("\r\n", "\n")
-        end
-
-        def raw(*command, &block)
-          int_callback = lambda do
-            @interrupted = true
-            @logger.info("Interrupted.")
-          end
-
-          # Append in the options for subprocess
-          command << { :notify => [:stdout, :stderr] }
-
-          Vagrant::Util::Busy.busy(int_callback) do
-            Vagrant::Util::Subprocess.execute(*command, &block)
-          end
+          @sudo_wrapper.run("lxc-#{command}", *args)
         end
       end
     end