+++ CHANGELOG.md
@@ -10,6 +10,7 @@ BACKWARDS INCOMPATIBILITIES:
 FEATURES:
 
   - Add support for salt-minion and add latest dev release for ubuntu codenamed saucy [#116](https://github.com/fgrehm/vagrant-lxc/pull/116)
+  - Add support for using a sudo wrapper script [#90](https://github.com/fgrehm/vagrant-lxc/issues/90)
 
 IMPROVEMENTS: